Security checks for the assessment application (+ some small related tweaks)

+++ dokeos/application/lib/assessment/assessment_manager/component/deleter.class.php
@@ -30,10 +30,17 @@ class AssessmentManagerDeleterComponent extends AssessmentManagerComponent
 			foreach ($ids as $id)
 			{
 				$assessment_publication = $this->retrieve_assessment_publication($id);
-
-				if (!$assessment_publication->delete())
+				
+				if (!$assessment_publication->is_visible_for_target_user($this->get_user()))
 				{
 					$failures++;
+				}
+				else
+				{
+					if (!$assessment_publication->delete())
+					{
+						$failures++;
+					}
 				}
 			}
